requirements.txt
requests
bs4
markdownify
pyfunctional